Scala Stream Collector: turn BufferConfig's byteLimit and recordLimit into longs (closes #3807)

+++ 2-collectors/scala-stream-collector/core/src/main/scala/com.snowplowanalytics.snowplow.collectors.scalastream/model.scala
@@ -102,7 +102,7 @@ package model {
   final case class Kafka(brokers: String, retries: Int) extends SinkConfig
   final case class Nsq(host: String, port: Int) extends SinkConfig
   case object Stdout extends SinkConfig
-  final case class BufferConfig(byteLimit: Int, recordLimit: Int, timeLimit: Long)
+  final case class BufferConfig(byteLimit: Long, recordLimit: Long, timeLimit: Long)
   final case class StreamsConfig(
     good: String,
     bad: String,
